Switch to using device slots
This removes the use of device IDs and instead identifies devices with
three-part PCI slot identifiers. A PCI slot identifier is a string of
the form:

```
bus:slot:function
```

Where `bus` is a value in the range `[0,255]`, `slot` is a value in
the range `[0,31]`, and `function` is a value in the range `[0,7]`.