Execute ifconfig commands to set up tap/vmnet devices
This execute various ifconfig commands prior to running a virtual
machine. Firstly, the tap/vmnet device is created with "ifconfig
create".  This step is allowed to fail; the device may already
exists. Secondly, the MAC address of the device is configured. This
step is not allowed to fail; if it fails, the device probably does
not exist. Lastly, the device is added to all of the groups defined,
if any are defined. These steps aren't allowed to fail.

Affects: https://github.com/io7m/waxmill/issues/24