Heavily refactor ZFS filesystem handling
This switches to a model where the waxmill package is assigned a
virtual machine runtime filesystem and takes complete ownership
of it. It knows the underlying ZFS filesystem name, and where the
filesystem is mounted. This ensures that any code that has to deal
with ZFS volumes automatically knows the filesystem that owns them,
and how to find them in /dev/zvol.

Fix: https://github.com/io7m/waxmill/issues/29