Use short IDs when communicating with Bhyve
BHyve has annoyingly small name length limits for virtual machines.
The name limits are not large enough to contain UUIDs in standard
encoding. This switches to using a shortened encoding of UUIDs
(URL-safe Base64) when communicating with bhyve, and adds a vm-id-of
command that can return IDs in long or short form.

Fix: https://github.com/io7m/waxmill/issues/30