Refactor network device handling
This refactors network device handling to guarantee that the host and
guest sides of a network device are assigned different MAC addresses.
This required replacing the --backend parameter to the command line
with a set of parameters. The information involved is too complex to
specify with a single command-line parameter.

This also updates the schema to enforce unique MAC addresses within
a virtual machine.

Fix: https://github.com/io7m/waxmill/issues/36